registering user and debugging

+++ src/utils/cloudinary.js
@@ -1,27 +1,35 @@
 import { v2 as cloudinary } from "cloudinary";
-import fs from "fs"
+import fs from "fs";
 
-(async function () {
-  cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET, 
-  });
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
-    try {
-        if (!localFilePath) return null 
-        const response = await cloudinary.uploader.upload(localFilePath , {
-            resource_type: "auto"
-        })
-        console.log("file is uploaded to cloudinary" , response.url);
-        return response
-
-    } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null;
+  try {
+    if (!localFilePath) return null;
+    
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
+    });
+    console.log("File uploaded to Cloudinary:", response.url);
+    
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-}
 
-export { uploadOnCloudinary }
+    return response;
+  } catch (error) {
+    console.error("Error uploading to Cloudinary:", error);
+    
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    
+    return null;
+  }
+};
+
+export { uploadOnCloudinary };
